Add quote show endpoint
