--- app/models/example.js
@@ -1,10 +1,0 @@
-const mongoose = require('mongoose');
-
-const ExampleSchema = new mongoose.Schema(
-  {
-    title: String
-  }
-);
-
-module.exports =  mongoose.model('Example', ExampleSchema);
-
