+++ app/models/quote.js
@@ -8,4 +8,3 @@ const QuoteSchema = new mongoose.Schema(
 );
 
 module.exports =  mongoose.model('Quote', QuoteSchema);
-
