--- config/routes/examples.js
@@ -1,8 +1,0 @@
-const express = require('express');
-const router = express.Router();
-const examplesController = require('../../app/controllers/examplesController');
-
-router.get('/', examplesController.index);
-router.get('/:id', examplesController.show);
-
-module.exports = router;
